Player Health(aka Armor) working
Health (armor bar) goes down every time player is shot.  I suggest
moving the enemies to a different position so that player doesn't get
hit when begins game

+++ cs113Project/Assets/Enemies/EnemiesShoots/EnemiesBullets.cs
@@ -4,12 +4,12 @@ using System.Collections;
 public class EnemiesBullets : MonoBehaviour {
 
 	//public int speed;
-	public int damage;
+	public float damage;
 	
 	// Use this for initialization
 	void Start () {
 		//speed = 100;
-		damage = 1;
+		damage = 1.0f;
 	}
 	
 	// Update is called once per frame
